Move common task-list functions into a standalone router
Adding task-lists to internal user's dashboard needs some, but not all of the same functionality as the external user dashboard. Abstract this into a single router that can be used in both instances.
