--- pages/dashboard/schema/index.js
@@ -1,18 +1,0 @@
-module.exports = {
-  updatedAt: {
-    show: true
-  },
-  establishment: {
-    show: true,
-    sortable: false
-  },
-  licence: {
-    show: true,
-    sortable: false
-  },
-  type: {
-    show: true,
-    sortable: false,
-    accessor: 'action.label'
-  }
-};
